update precision to IPL_DEPTH_32F bug fixes in psiDerivative adding toolsKit::costumeLineCompute removing downscale in getDXsCVSobel removing upscale in computePsidashFS_brox ans1&ans2 removing downscale back in each iter in SmoothFlowPDE

+++ trunk/constructMatrix_brox.h
@@ -10,7 +10,7 @@ class constructMatrix_brox
 public:
 	constructMatrix_brox(void);
 	static void constructMatrix_b(IplImage* Ikx,IplImage* Iky,IplImage* Ikz,IplImage* Ixx,IplImage* Ixy,IplImage* Iyy,IplImage* Ixz,
-								  IplImage* Iyz,IplImage* psidash,IplImage* psidashFS1,IplImage* psidashFS2,IplImage*  u,IplImage*  v,double gamma,int _ERROR_CONST );
+								  IplImage* Iyz,IplImage* psidash,IplImage* psidashFS1,IplImage* psidashFS2,IplImage*  u,IplImage*  v,double gamma,double _ERROR_CONST );
 	virtual ~constructMatrix_brox(void);
 private:
